Теперь пули уничтожаются при столкновении с препятствием. Проведён полный рефакторинг кода
