+++ characters/Kiana/kiana.py
@@ -1,0 +1,41 @@
+import pygame
+import sys
+import os
+
+
+class Kiana(pygame.sprite.Sprite):
+    def __init__(self, *group, fps):
+        super().__init__(*group)
+        self.frames = [f"Kiana{i}.png" for i in range(2)]
+        self.fps = fps
+        self.cur_frame = 0
+        self.image = self.load_image(self.frames[self.cur_frame])
+        self.image = pygame.transform.scale(self.image, (80, 80))
+        self.rect = self.image.get_rect()
+        self.rect.x, self.rect.y = 660, 360
+        self.clock = 0
+
+    def update(self):
+        if self.clock == 2500 // self.fps:
+            self.clock = 0
+            self.cur_frame = (self.cur_frame + 1) % len(self.frames)
+            self.image = self.load_image(self.frames[self.cur_frame])
+            self.image = pygame.transform.scale(self.image, (80, 80))
+        else:
+            self.clock += 1
+
+    def load_image(self, name, colorkey=None):
+        fullname = os.path.join('images/characters/Kiana', name)
+        if not os.path.isfile(fullname):
+            print(f"Файл с изображением '{fullname}' не найден")
+            sys.exit()
+        image = pygame.image.load(fullname)
+
+        if colorkey is not None:
+            image = image.convert()
+            if colorkey == -1:
+                colorkey = image.get_at((0, 0))
+            image.set_colorkey(colorkey)
+        else:
+            image = image.convert_alpha()
+        return image
