+++ images/characters/Kiana/bullet.png
